Move WebsocketServer under the module RemoteConsole as RackServer
